backend/functions/requirements.txt
functions-framework>=3.5.0
google-cloud-storage>=2.10.0
google-cloud-aiplatform>=1.38.0
google-cloud-speech>=2.22.0